+++ tests/security/http/test_security_api_key_cookie.py
@@ -54,7 +54,7 @@ def test_security_api_key_no_key():
         ],
     ) as client:
         response = client.get("/users/me")
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.text == "Not authenticated"