+++ tests/security/http/test_security_api_key_with_requires_injects.py
@@ -62,7 +62,7 @@ def test_security_api_key_no_key():
         middleware=[DefineMiddleware(RequestContextMiddleware)],
     ) as client:
         response = client.get("/users/me")
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.text == "Not authenticated"