+++ tests/security/http/test_security_http_base.py
@@ -43,10 +43,10 @@ def test_security_http_base_no_credentials():
                 handler=read_current_user,
                 dependencies={"credentials": Provide(security)},
             )
-        ]
+        ],
     ) as client:
         response = client.get("/users/me")
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.text == "Not authenticated"