+++ tests/security/openid/test_security_openid_connect_description.py
@@ -72,7 +72,7 @@ def test_security_oauth2_password_bearer_no_header():
         enable_openapi=True,
     ) as client:
         response = client.get("/users/me")
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.text == "Not authenticated"